# Change highlighting to not include duplicated background colors