Fixes #415