@Keats 2018-09-09
Can you run the benchmarks of the rendering crate before/after to see if there are any differences? The tests failing are easily fixable and not really good actually.