@porglezomp 2018-09-09
Rebased onto next.
Here's my benchmark results.

Before:
```
test bench_render_content_no_shortcode         ... bench:      47,583 ns/iter (+/- 6,082)
test bench_render_content_with_highlighting    ... bench:   1,047,409 ns/iter (+/- 219,777)
test bench_render_content_without_highlighting ... bench:     165,437 ns/iter (+/- 34,182)
test bench_render_shortcodes_one_present       ... bench:     109,179 ns/iter (+/- 90,657)
```

After:
```
test bench_render_content_no_shortcode         ... bench:      49,520 ns/iter (+/- 22,794)
test bench_render_content_with_highlighting    ... bench:     946,633 ns/iter (+/- 149,835)
test bench_render_content_without_highlighting ... bench:     174,963 ns/iter (+/- 53,546)
test bench_render_shortcodes_one_present       ... bench:     107,826 ns/iter (+/- 17,610)
```

Doesn't look like any significant change.